Change max_pending variable type to int32_t.
Fixes ticket #2829.

git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@7617 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ globals.h
@@ -1649,7 +1649,7 @@ struct s_config
 
 	char		*ecmfmt;
 
-	uint8_t max_pending;
+	int32_t		max_pending;
 };
 
 struct s_clientinit
